:zap: feat: modify ID and add out method

+++ honey-back/ggulswriting/src/main/java/com/thirty/ggulswriting/member/repository/MemberRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.thirty.ggulswriting.member.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Integer> {
-	Optional<Member> findMemberByMemberIdAndGoodbyeTimeIsNull(Long memberId);
+	Optional<Member> findMemberByMemberIdAndGoodbyeTimeIsNull(int memberId);
 }
